requirements.txt
pytest
pyserial-asyncio